Reorganize source code
