+++ android/app/build.gradle
@@ -16,7 +16,7 @@ android {
         minSdk 26
         targetSdk 34
 
-        applicationId "com.gta.reversed"
+        applicationId "com.regta.core"
         versionName "reGTA-0.0.1"
         versionCode 1
 
@@ -49,7 +49,7 @@ android {
 
     ndkVersion '26.3.11579264'
 
-    namespace 'com.gta.reversed'
+    namespace 'com.regta.core'
 
     buildTypes {
         release {
