+++ android/app/src/main/AndroidManifest.xml
@@ -44,7 +44,7 @@
        >
 
         <activity
-            android:name="com.gta.launcher.MainActivity"
+            android:name="com.regta.launcher.MainActivity"
             android:exported="true"
             android:screenOrientation="landscape"
             android:theme="@style/Theme.AppCompat.Light.NoActionBar.FullScreen">
@@ -55,7 +55,7 @@
         </activity>
 
         <activity
-            android:name="com.gta.reversed.REGTA"
+            android:name="com.regta.core.REGTA"
             android:exported="false"
             android:launchMode="singleTask"
             android:screenOrientation="landscape"
