+++ android/app/src/main/cpp/src/StdInc.cpp
@@ -1,5 +1,5 @@
 //
-// Created by unk.cs on 19/01/2025.
+// Created by SyXhOwN on 19/01/2025.
 //
 
 #include "StdInc.h"
