+++ android/app/src/main/cpp/src/StdInc.h
@@ -1,16 +1,5 @@
 //
-// Created by unk.cs on 19/01/2025.
+// Created by SyXhOwN on 19/01/2025.
 //
 
-#ifndef REGTA_STDINC_H
-#define REGTA_STDINC_H
-
-
-
-class StdInc {
-
-};
-
-
-
-#endif //REGTA_STDINC_H
+#pragma once
