+++ android/app/src/main/cpp/src/reGTA.cpp
@@ -2,12 +2,22 @@
 // Created by SyXhOwN on 17/01/2025.
 //
 
-#include "main.h"
+#include "reGTA.h"
+
+// JavaWrapper
+#include "Vendor/Javawrapper/JavaWrapper.h"
+
+// ReversibleHooks
+#include "Vendor/ReversibleHooks/ReversibleHooks.h"
+
+// Log
+#include "Vendor/Log/Log.h"
 
 // Trace
 #include "Vendor/Trace/StackTrace.h"
 
 void InstallHooks();
+void InjectReversedClasses();
 
 uintptr_t g_libGTASA = 0;
 uintptr_t g_libREGTA = 0;
@@ -26,7 +36,7 @@ void PrintBuildCrashInfo()
     CrashLog("Build times: %s %s. ABI: %s", __TIME__, __DATE__, (VERSION_x32 ? "armeabi-v7a" : "arm64-v8a"));
 }
 
-void handler(int signum, siginfo_t *info, void* contextPtr)
+void Handler(int signum, siginfo_t *info, void* contextPtr)
 {
     ucontext* context = (ucontext_t*)contextPtr;
 
@@ -48,7 +58,7 @@ void handler(int signum, siginfo_t *info, void* contextPtr)
     }
 }
 
-void handler1(int signum, siginfo_t* info, void* contextPtr)
+void Handler1(int signum, siginfo_t* info, void* contextPtr)
 {
     auto context = (ucontext_t*)contextPtr;
 
@@ -70,7 +80,7 @@ void handler1(int signum, siginfo_t* info, void* contextPtr)
     }
 }
 
-void handler2(int signum, siginfo_t* info, void* contextPtr)
+void Handler2(int signum, siginfo_t* info, void* contextPtr)
 {
     ucontext* context = (ucontext_t*)contextPtr;
 
@@ -93,7 +103,7 @@ void handler2(int signum, siginfo_t* info, void* contextPtr)
 
 }
 
-void handler3(int signum, siginfo_t* info, void* contextPtr)
+void Handler3(int signum, siginfo_t* info, void* contextPtr)
 {
     ucontext* context = (ucontext_t*)contextPtr;
 
@@ -129,7 +139,7 @@ JNIEXPORT jint JNI_OnLoad(JavaVM *vm, void *reserved)
 {
     javaVM = vm;
 
-    g_libGTASA = Patch::FindLib("libGTASA.so");
+    g_libGTASA = ReversibleHooks::FindLib("libGTASA.so");
 
     if(g_libGTASA == 0) {
         Log("ERROR: libGTASA.so address not found!");
@@ -138,7 +148,7 @@ JNIEXPORT jint JNI_OnLoad(JavaVM *vm, void *reserved)
 
     Log("libGTASA.so image base address: 0x%X", g_libGTASA);
 
-    g_libREGTA = Patch::FindLib("libreGTA.so");
+    g_libREGTA = ReversibleHooks::FindLib("libreGTA.so");
 
     if(g_libREGTA == 0) {
         Log("ERROR: libreGTA.so address not found!");
@@ -147,30 +157,33 @@ JNIEXPORT jint JNI_OnLoad(JavaVM *vm, void *reserved)
 
     Log("libreGTA.so image base address: 0x%X", g_libREGTA);
 
-    Patch::InitHookStuff();
+    ReversibleHooks::InitHookStuff();
 
     InstallHooks();
 
+    // TODO: Activate the function below for the inverted classes to be replaced.
+    // InjectReversedClasses();
+
     struct sigaction act;
-    act.sa_sigaction = handler;
+    act.sa_sigaction = Handler;
     sigemptyset(&act.sa_mask);
     act.sa_flags = SA_SIGINFO;
     sigaction(SIGSEGV, &act, &act_old);
 
     struct sigaction act1;
-    act1.sa_sigaction = handler1;
+    act1.sa_sigaction = Handler1;
     sigemptyset(&act1.sa_mask);
     act1.sa_flags = SA_SIGINFO;
     sigaction(SIGABRT, &act1, &act1_old);
 
     struct sigaction act2;
-    act2.sa_sigaction = handler2;
+    act2.sa_sigaction = Handler2;
     sigemptyset(&act2.sa_mask);
     act2.sa_flags = SA_SIGINFO;
     sigaction(SIGFPE, &act2, &act2_old);
 
     struct sigaction act3;
-    act3.sa_sigaction = handler3;
+    act3.sa_sigaction = Handler3;
     sigemptyset(&act3.sa_mask);
     act3.sa_flags = SA_SIGINFO;
     sigaction(SIGBUS, &act3, &act3_old);
@@ -180,9 +193,9 @@ JNIEXPORT jint JNI_OnLoad(JavaVM *vm, void *reserved)
 
 extern "C"
 JNIEXPORT void JNICALL
-Java_com_gta_reversed_REGTA_initreGTA(JNIEnv *env, jobject thiz)
+Java_com_regta_core_REGTA_initreGTA(JNIEnv *env, jobject initreGTA)
 {
     Log("Initializing reGTA..");
 
-    g_pJavaWrapper = new CJavaWrapper(env, thiz);
+    g_pJavaWrapper = new CJavaWrapper(env, initreGTA);
 }
